Move rest.Pods to pkg/registry/pod/rest.go
